feat: 将选择生日和选择职业组件化
